Change mapping from using expression-quote syntax to using raw strings.
This fixes #6. This change should allow Mac keys like Cmd-D to be mapped
correctly. Previously, keys are mapped using the expression quote syntax, which
gets turned into raw key code in the map command. The keycode makes it easy to
check for input when doing getchar() in multicursor mode, but appears to cause
problems with the noremap command, and is really not a standard way to map keys
anyway. This change now uses raw strings, and convert the raw string to a
keycode at runtime (using some voodoo exec magic)