Add support for searching whole words
VIM's * command searches for the whole word -- it wraps current word with word
boundaries: \< and \>. g*, on the other hand, does not care about word
boundaries.

This patch adds functionality to configure multiple cursors behaviour to be
consistent with VIM's * and g*. To do that, use newly introduced option:

    let g:multi_cursor_start_key='g<C-n>'
    let g:multi_cursor_start_word_key='<C-n>'

Note that in current implementation "start word" command works in both Normal
and Visual modes.