Update setting default in documentation
Documented the new default for `g:multi_cursor_normal_maps`

I removed the spaces after the commas in the value, because Vim doesn’t highlight helpfile text in backticks if it contains spaces.

I couldn’t think of any obvious way to write this long value readably without breaking up the flow of the text too much, so I just let it go off the edge of the screen.