Revise docs for g:multi_cursor_normal_maps
I make it clear that the default value should work for most people, and describe when people will want to customize the value.

I also made some markup more consistent between the two versions so they are easier to just copy and paste when one version is edited.

And I added spaces to the default value, because I think it’s more important to make the keys of the map clear than it is to make the map shorter, now that it is written on its own line.