Fix get_time_in_ms() type; reintroduce sleep and insert_map tests
Removing the sleep's was causing a 100% CPU usage for insert completion
and special key completion.
The reason why it got removed in the first place was because it can
produce misleading timings (especially on slow system).
The workaround is to sleep for long enough (100ms) at each call to
avoid sleep's call overhead impacting the total time too much.